+++ src/ios/OSInAppBrowserInputArgumentsModel.swift
@@ -1,4 +1,3 @@
-import OSInAppBrowserLib
 
 class OSInAppBrowserInputArgumentsSimpleModel: Decodable {
     let url: String